+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.4.2
+
+### Module Migration
+
+* Late `@import` rules can now be hoisted above safe at rules in the same file.
+
+* Fix a typo in the error message for late `@import` rules that could not be
+  migrated.
+
+* Fix a bug where hoisted `@use` rules with configuration would have incorrect
+  syntax.
+
 ## 2.4.1
 
 ### Module Migrator